// apps/website/src/lib/types.ts
/* -------------------------------------------------------------------

                ⚡ Storm Software - Flash Flood

 This code was released as part of the Flash Flood project. Flash Flood
 is maintained by Storm Software under the Apache-2.0 License, and is
 free for commercial and private use. For more information, please visit
 our licensing page.

 Website:         https://stormsoftware.com
 Repository:      https://github.com/storm-software/flash-flood
 Documentation:   https://stormsoftware.com/projects/flash-flood/docs
 Contact:         https://stormsoftware.com/contact
 License:         https://stormsoftware.com/projects/flash-flood/license

 ------------------------------------------------------------------- */

export type ServerActionResult<TData> =
  | {
      success: true;
      data?: TData;
    }
  | {
      success: false;
      error: string;
      errors?: Record<string, string[] | undefined>;
    };

export type BlockchainType = "ethereum" | "solana";